+++ dapp/components/dashboard/nft-area/nft-area.tsx
@@ -4,7 +4,7 @@ import classNames from 'classnames';
 import { useState } from 'react';
 import { ethers } from 'ethers';
 import { useAddressData, useAddressNfts } from '../../../hooks/address-data';
-import { NftItem, FetchNftDataResponse } from '../../../types/types';
+import { NftItem, FetchNftDataResponse } from '../../../types/typesNftApi';
 import { fetchAddressData } from '@/utils/frontendUtils';
 import { BigNumber } from 'ethers';
