+++ dapp/tsconfig.json
@@ -17,6 +17,6 @@
       "@/*": ["./*"]
     }
   },
-  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", "types/types.d.ts"],
+  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", "types/typesNftApi.d.ts"],
   "exclude": ["node_modules"]
 }
